add TodoContainer component, code refactoring
