+++ src/components/TodoContainer.module.css
@@ -1,0 +1,11 @@
+.Section {
+  width: 50vw;
+  margin: 0 auto;
+  margin-top: 8rem;
+  max-width: 30rem;
+  background: #fff;
+  border-radius: 0.25rem;
+  padding: 2rem;
+  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.1), 0 1px 2px 0 rgba(0, 0, 0, 0.06);
+  transition: 0.3s ease-in-out all;
+}
